acl_ivan_bot v0.0.1 Alice.Handlers.Standup

Alice.Handlers.Standup is a slack handler for managing incoming slack bot messages

Functions

commands()

All of the commands handled by this module

daily_report(conn)

delete_state(conn, key)

Deletes the entries in the state for a specific key.

get_state(conn, key, default \\ nil)

Get the state from an Alice.Conn struct, namespaced to this module

guide(conn)

match_commands(conn)

Match all commands in this module

match_routes(conn)

Match all routes in this module

put_state(conn, key, value)

Update the state of an Alice.Conn struct, namespaced to this module

routes()

All of the routes handled by this module

standup(conn)

who_cares(conn)

yesterday_report(conn)